+++ lib/src/commands/commands.dart
@@ -1,4 +1,5 @@
 export 'add_locale_command.dart';
+export 'bottom_sheet_command.dart';
 export 'change_app_name_command.dart';
 export 'change_id_command.dart';
 export 'create_command.dart';