+++ lib/src/templates/run_config_templates.dart
@@ -1,29 +1,26 @@
-const runConfigDevelopmentTemplate = r'''
-<component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="development" type="FlutterRunConfigurationType" factoryName="Flutter">
-    <option name="buildFlavor" value="development" />
-    <option name="filePath" value="$PROJECT_DIR$/lib/main_development.dart" />
-    <method v="2" />
-  </configuration>
-</component>
-''';
+const runConfigDevelopmentTemplate =
+    '<component name="ProjectRunConfigurationManager">\n'
+    '  <configuration default="false" name="development" type="FlutterRunConfigurationType" factoryName="Flutter">\n'
+    '    <option name="buildFlavor" value="development" />\n'
+    '    <option name="filePath" value="\$PROJECT_DIR\$/lib/main_development.dart" />\n'
+    '    <method v="2" />\n'
+    '  </configuration>\n'
+    '</component>\n';
 
-const runConfigStagingTemplate = r'''
-<component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="staging" type="FlutterRunConfigurationType" factoryName="Flutter">
-    <option name="buildFlavor" value="staging" />
-    <option name="filePath" value="$PROJECT_DIR$/lib/main_staging.dart" />
-    <method v="2" />
-  </configuration>
-</component>
-''';
+const runConfigStagingTemplate =
+    '<component name="ProjectRunConfigurationManager">\n'
+    '  <configuration default="false" name="staging" type="FlutterRunConfigurationType" factoryName="Flutter">\n'
+    '    <option name="buildFlavor" value="staging" />\n'
+    '    <option name="filePath" value="\$PROJECT_DIR\$/lib/main_staging.dart" />\n'
+    '    <method v="2" />\n'
+    '  </configuration>\n'
+    '</component>\n';
 
-const runConfigProductionTemplate = r'''
-<component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="production" type="FlutterRunConfigurationType" factoryName="Flutter">
-    <option name="buildFlavor" value="production" />
-    <option name="filePath" value="$PROJECT_DIR$/lib/main_production.dart" />
-    <method v="2" />
-  </configuration>
-</component>
-''';
+const runConfigProductionTemplate =
+    '<component name="ProjectRunConfigurationManager">\n'
+    '  <configuration default="false" name="production" type="FlutterRunConfigurationType" factoryName="Flutter">\n'
+    '    <option name="buildFlavor" value="production" />\n'
+    '    <option name="filePath" value="\$PROJECT_DIR\$/lib/main_production.dart" />\n'
+    '    <method v="2" />\n'
+    '  </configuration>\n'
+    '</component>\n';